% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_quasi_murioshi.R
\name{generate_quasi_murioshi}
\alias{generate_quasi_murioshi}
\title{Generating quasi murioshi of popn music}
\usage{
generate_quasi_murioshi()
}
\description{
Generating quasi murioshi of popn music
}
